backend/requirements.txt
fastapi==0.136.0
# Starlette is a transitive dep via fastapi; pin floor to mitigate
# GHSA-f96h-pmfr-66vw and GHSA-2c2j-9gv5-cj73 (both fixed in 0.40.0 / 0.47.2).
starlette>=0.48.0
uvicorn[standard]==0.32.0
watchdog==5.0.3
psutil==6.1.0
sse-starlette==2.1.3
pydantic==2.9.2
# Native desktop window around the React UI (Edge WebView2 on Windows,
# WebKit on macOS, WebKitGTK on Linux).
pywebview==5.3.2
# Windows-only — used for per-tab UI Automation focus. Skipped on Linux/mac (CI).
uiautomation==2.0.29; platform_system == "Windows"
comtypes==1.4.11; platform_system == "Windows"
# PTY backend for the embedded terminal. Windows uses pywinpty (ConPTY);
# posix platforms use ptyprocess. pywinpty is PINNED to 2.0.15 — 3.x
# regresses in PyInstaller `console=False` frozen exes: cmd.exe exits
# with STATUS_CONTROL_C_EXIT (0xC000013A) within ms of spawn. 2.0.15
# works correctly in both dev and frozen contexts.
pywinpty==2.0.15; platform_system == "Windows"
ptyprocess==0.7.0; platform_system != "Windows"
